requirements.txt
flask
pandas
groq
python-dotenv
